Chose micro data center where to allocate VM in a greedy manner.

+++ src/main/java/cloud/cluster/sim/clustersimulator/dto/Chance.java
@@ -27,7 +27,7 @@ public class Chance {
     }
 
     public double getChance() {
-        chance = favorable / possible;
+        chance = (double)favorable / possible;
         return chance;
     }
